server/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib
python-multipart
sqlalchemy
python-dotenv
psycopg2
bcrypt